Separate NewCategorySheetView into its own file
