+++ brain-marks/Categories/Views/CategoryList.swift
@@ -48,54 +48,8 @@ struct CategoryList: View {
             viewModel.getCategories()
         }
         .accentColor(.black)
-                .sheet(isPresented:$showAddURLView) {
-                    AddURLView(categories: viewModel.categories)
-                }
-    }
-}
-
-struct NewCategorySheetView: View {
-    @Environment(\.presentationMode) var presentationMode
-    @State var newCategory = ""
-    
-    var body: some View {
-        NavigationView {
-            VStack {
-                
-                TextField("Enter name of new category", text: $newCategory)
-                    .padding()
-                
-                Spacer()
-                
-                HStack(spacing: 25) {
-                    Button {
-                        presentationMode.wrappedValue.dismiss()
-                    } label: {
-                        Text("Cancel")
-                            .frame(width: 150, height: 50)
-                            .foregroundColor(.white)
-                            .background(Color(UIColor(named: "twitter")!))
-                            .font(.system(size: 20, weight: .semibold, design: .default))
-                            .cornerRadius(10)
-                    }
-                    
-                    Button {
-                        presentationMode.wrappedValue.dismiss()
-                        DataStoreManger.shared.createCategory(
-                            category: AWSCategory(name: newCategory,
-                                                  imageName: "swift"))
-                    } label: {
-                        Text("Create")
-                            .frame(width: 150, height: 50)
-                            .foregroundColor(.white)
-                            .background(Color(UIColor(named: "twitter")!))
-                            .font(.system(size: 20, weight: .semibold, design: .default))
-                            .cornerRadius(10)
-                    }
-                }
-                .padding(20)
-            }
-            .navigationBarTitle("New Category")
+        .sheet(isPresented:$showAddURLView) {
+            AddURLView(categories: viewModel.categories)
         }
     }
 }
@@ -103,7 +57,5 @@ struct NewCategorySheetView: View {
 struct CategoryList_Previews: PreviewProvider {
     static var previews: some View {
         CategoryList()
-        
-        NewCategorySheetView()
     }
 }
